Refactor MTE-2598 BasicBrowsing and TestRail (#20182)
* 🈳 Empty commit

* Rename BasicBrowsing to Browsing

* Oops. Should be BrowsingTest

* Enable RequestDesktopTest :tada:

* Move testActivityMenuRequestDesktopItem to BrowsingTest. Add TestRail link.

* Add TestRail URL to testCheckCollapsedURL

* Remove CollapsedURLTest file

* Remove CollapsedURLTest file

* Comment out known issue with the link
